+++ src/utils/index.ts
@@ -6,3 +6,4 @@ export * from "./debounce";
 export * from "./invariant";
 export * from "./batch";
 export * from "./selectors";
+export * from "./backoff";